conversation: extract anthropic_models.ts from anthropic.ts
